Split archives to their dedicated blueprint, refs #195
This way, we can fallback to archives routes if no other match is found for regular routes.
